Maintenance - Refactored Language Dropdown
What
=================
Refactored language dropdown to now be defined in functions.php and
called via `do_action('ct_language_dropdown');`. It uses WPML attributes
and is therefore dependent on WPML to being installed. It uses WPML
flags instead of those defined in the CSS, so those definitions in the
CSS for the flags have been deleted.

Why
=================
To make it integrated with WPML and centralised in functions.php whilst
also being flexible (now that it uses WPML) with the changes made in the
admin area in WPML

Ticket: #2 and #3