Enhancement - Added More Client Logos + Deleted unwanted Clients
What
=================
Removed and added the clients as required.
Refactored to iterate over the client files in the filesystem and build
the list items for the client sideways scroll dynamically rather than
placing the html manually.

I'm looking to completely remove the need to having the CSS dictate the
location of the file so that it will be completely dynamic depending
only on the images which are present in the "img" folder.

Beyond that, if there is still time, I could incorporate this whole
precess into the admin area of WordPress, but for now uploading image
files and then having the rest be done for you should be enough.

Ticket: #7