Maintenance - Refactored template elements
What
=================
Added a link to the phone number in the header. This revealed a failure
in the stylesheet to address a:hover, a:active, a:focus, which has been
corrected.
Some hard coded links were found and changed due to the fact that 1. it
was pulling assets in on an insecure connection (dev site is on https)
2. it was pulling assets from a different server/domain. These are now
dynamic using the wordpress functions as they should have been in the
first place.

Ticket: #2