Maintenance - Refactored Price Enquiries
What
=================
Major refactor.
Some fields were made optional, they were all required before.
References to the website were made dynamic (you never know where this
is going to be hosted).
Style fixes applied to make compatible with Grammarly (thanks for the
heads up Kristo!) since Grammarly was manipulating the styles of the
form and the textbox within.

Why
=================
Less required fields will mean a better user experience as some of these
fields require information not easily at hand for a general translations
seeker.
Many people will most likely be using Grammarly in the future if not
already, as it has become popular in the business world as well as with
the general public. This means when Grammarly does something like make
all the fields on your page disappear when the textarea is selected,
then we need to make sure the form is compatible with this addon.

Ticket: #2