Feature - Completed CT Settings Page
What
=================
Completed the settings page enabling the editing of company details by
the client.
Also needed to restructure the data into a single data entry instead of
3 separate ones due to settings page construction issues. The rest of
the site needs a refactor to rectify any bugs as a result.

Why
=================
So that the client is able to change their name, phone number, email,
website, etc. themselves.

Ticket: #8