Bugfix - Homepage background images weren't looping
What
=================
I changed the function reference into an actual function at the
beginning of this project, unaware that setInterval wouldn't like that.
That has been fixed now and I finally found time to implement a check to
stop the loop if on any page other than the homepage.

Why
=================
Images stopped their transitions after the first loop.

Ticket: N/A