Bufix - Client logo system not taking into account logo numbers
What
=================
The client logo loop wasn't taking the number of the actual file, it was
generating its on each iteration of the loop and using that. I've
changed it now to extract the image number and use that instead. As a
result, a missing file in the filesystem will no longer affect the other
client logos as it did in the past. It was applying a different class to
what it was supposed to to the client logo. This was fine, until a file
with a number in between the others was deleted. Then that class would
still be applied but wouldn't find the file (because it's missing...)
resulting in an empty list item taking up space as well as the client
logos near the end of the loop sometimes not being rendered due to
the loop thinking there are no more files to consider.

Why
=================
To ensure as automated an dynamic a system as possible. This way a
simple change in the files in the filesystem will affect what the loop
displays.
AKA the loop be shmart!

Ticket: #7